+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/ApiDuplicatorService_CreateWithDefinitionTest.java
@@ -22,28 +22,18 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Charsets;
 import com.google.common.io.Resources;
 import io.gravitee.repository.exceptions.TechnicalException;
-import io.gravitee.repository.management.api.ApiRepository;
-import io.gravitee.repository.management.api.MembershipRepository;
 import io.gravitee.repository.management.model.Api;
 import io.gravitee.rest.api.model.*;
 import io.gravitee.rest.api.model.api.ApiEntity;
-import io.gravitee.rest.api.model.parameters.Key;
-import io.gravitee.rest.api.model.parameters.ParameterReferenceType;
 import io.gravitee.rest.api.model.permissions.RoleScope;
-import io.gravitee.rest.api.model.permissions.SystemRole;
 import io.gravitee.rest.api.service.exceptions.PlanNotFoundException;
-import io.gravitee.rest.api.service.impl.ApiServiceImpl;
-import io.gravitee.rest.api.service.notification.NotificationTemplateService;
-import io.gravitee.rest.api.service.search.SearchEngineService;
+import io.gravitee.rest.api.service.impl.ApiDuplicatorServiceImpl;
 import io.gravitee.rest.api.service.spring.ServiceConfiguration;
 import java.io.IOException;
-import java.io.Reader;
 import java.net.URL;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Optional;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
@@ -60,26 +50,20 @@ import org.springframework.security.core.context.SecurityContextHolder;
  * @author GraviteeSource Team
  */
 @RunWith(MockitoJUnitRunner.class)
-public class ApiService_CreateWithDefinitionTest {
+public class ApiDuplicatorService_CreateWithDefinitionTest {
 
     private static final String API_ID = "id-api";
     private static final String PLAN_ID = "my-plan";
     private static final String SOURCE = "source";
 
     @InjectMocks
-    private ApiServiceImpl apiService = new ApiServiceImpl();
-
-    @Mock
-    private ApiRepository apiRepository;
-
-    @Mock
-    private MembershipRepository membershipRepository;
+    protected ApiDuplicatorService apiDuplicatorService = new ApiDuplicatorServiceImpl();
 
     @Spy
     private ObjectMapper objectMapper = (new ServiceConfiguration()).objectMapper();
 
     @Mock
-    private Api api;
+    private ApiService apiService;
 
     @Mock
     private MembershipService membershipService;
@@ -100,34 +84,7 @@ public class ApiService_CreateWithDefinitionTest {
     private RoleService roleService;
 
     @Mock
-    private AuditService auditService;
-
-    @Mock
-    private IdentityService identityService;
-
-    @Mock
-    private SearchEngineService searchEngineService;
-
-    @Mock
-    private ParameterService parameterService;
-
-    @Mock
-    private VirtualHostService virtualHostService;
-
-    @Mock
-    private GenericNotificationConfigService genericNotificationConfigService;
-
-    @Mock
     private ApiMetadataService apiMetadataService;
-
-    @Mock
-    private AlertService alertService;
-
-    @Spy
-    private PolicyService policyService;
-
-    @Mock
-    private NotificationTemplateService notificationTemplateService;
 
     @AfterClass
     public static void cleanSecurityContextHolder() {
@@ -145,14 +102,6 @@ public class ApiService_CreateWithDefinitionTest {
         );
     }
 
-    @Before
-    public void setUp() {
-        when(notificationTemplateService.resolveInlineTemplateWithParam(anyString(), any(Reader.class), any()))
-            .thenReturn("toDecode=decoded-value");
-        when(parameterService.find(Key.API_PRIMARY_OWNER_MODE, ParameterReferenceType.ENVIRONMENT)).thenReturn("USER");
-        reset(searchEngineService);
-    }
-
     @Test
     public void shouldCreateImportApiWithMembersAndPages() throws IOException, TechnicalException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition+members+pages.json");
@@ -161,8 +110,7 @@ public class ApiService_CreateWithDefinitionTest {
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
         when(userService.findBySource(anyString(), anyString(), eq(false))).thenReturn(new UserEntity());
         RoleEntity poRoleEntity = new RoleEntity();
         poRoleEntity.setId("API_PRIMARY_OWNER");
@@ -191,31 +139,19 @@ public class ApiService_CreateWithDefinitionTest {
         user.setSource(SOURCE);
         user.setSourceId("ref-user");
         when(userService.findBySource(user.getSource(), user.getSourceId(), false)).thenReturn(user);
-        when(userService.findById(admin.getId())).thenReturn(admin);
         MemberEntity memberEntity = new MemberEntity();
         memberEntity.setId(admin.getId());
         memberEntity.setRoles(Collections.singletonList(poRoleEntity));
-        when(membershipService.addRoleToMemberOnReference(any(), any(), any())).thenReturn(memberEntity);
-        when(membershipService.addRoleToMemberOnReference(any(), any(), any())).thenReturn(memberEntity);
         when(userService.findBySource(user.getSource(), user.getSourceId(), false)).thenReturn(user);
         when(userService.findById(memberEntity.getId())).thenReturn(admin);
         when(pageService.createWithDefinition(any(), any())).thenReturn(new PageEntity());
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
         verify(pageService, times(2)).createWithDefinition(eq(API_ID), anyString());
         verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember(admin.getId(), null, MembershipMemberType.USER),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
-        verify(membershipService, times(1))
             .addRoleToMemberOnReference(MembershipReferenceType.API, API_ID, MembershipMemberType.USER, user.getId(), "API_OWNER");
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 
     @Test
@@ -226,8 +162,8 @@ public class ApiService_CreateWithDefinitionTest {
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
@@ -238,7 +174,6 @@ public class ApiService_CreateWithDefinitionTest {
         user.setSourceId("ref-user");
         when(userService.findBySource(user.getSource(), user.getSourceId(), false)).thenReturn(user);
         when(userService.findById(admin.getId())).thenReturn(admin);
-        when(groupService.findByEvent(any())).thenReturn(Collections.emptySet());
 
         RoleEntity poRoleEntity = new RoleEntity();
         poRoleEntity.setId("API_PRIMARY_OWNER");
@@ -261,24 +196,14 @@ public class ApiService_CreateWithDefinitionTest {
         MemberEntity memberEntity = new MemberEntity();
         memberEntity.setId(admin.getId());
         memberEntity.setRoles(Collections.singletonList(poRoleEntity));
-        when(membershipService.addRoleToMemberOnReference(any(), any(), any())).thenReturn(memberEntity);
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
         verify(pageService, times(1)).createPage(eq(API_ID), any(NewPageEntity.class));
         verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember(admin.getId(), null, MembershipMemberType.USER),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
-        verify(membershipService, times(1))
             .addRoleToMemberOnReference(MembershipReferenceType.API, API_ID, MembershipMemberType.USER, user.getId(), "API_OWNER");
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
         verify(membershipService, never()).transferApiOwnership(any(), any(), any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 
     @Test
@@ -289,8 +214,8 @@ public class ApiService_CreateWithDefinitionTest {
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
@@ -299,28 +224,24 @@ public class ApiService_CreateWithDefinitionTest {
         user.setId("user");
         user.setSource(SOURCE);
         user.setSourceId("ref-user");
-        when(userService.findById(admin.getId())).thenReturn(admin);
         when(pageService.createWithDefinition(any(), any())).thenReturn(new PageEntity());
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
         verify(pageService, times(2)).createWithDefinition(eq(API_ID), anyString());
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 
     @Test
-    public void shouldCreateImportApiWithOnlyDefinition() throws IOException, TechnicalException {
+    public void shouldCreateImportApiWithOnlyDefinition() throws IOException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition.json");
         String toBeImport = Resources.toString(url, Charsets.UTF_8);
         ApiEntity apiEntity = new ApiEntity();
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
@@ -329,33 +250,23 @@ public class ApiService_CreateWithDefinitionTest {
         user.setId("user");
         user.setSource(SOURCE);
         user.setSourceId(API_ID);
-        when(userService.findById(admin.getId())).thenReturn(admin);
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
-        verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember("admin", null, MembershipMemberType.USER),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
     }
 
     @Test
-    public void shouldCreateImportApiWithOnlyDefinitionWithPrimaryOwner() throws IOException, TechnicalException {
+    public void shouldCreateImportApiWithOnlyDefinitionWithPrimaryOwner() throws IOException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition+primaryOwner.json");
         String toBeImport = Resources.toString(url, Charsets.UTF_8);
         ApiEntity apiEntity = new ApiEntity();
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
@@ -365,77 +276,22 @@ public class ApiService_CreateWithDefinitionTest {
         user.setSource(SOURCE);
         user.setSourceId(API_ID);
 
-        when(userService.findById(any())).thenReturn(user);
-        when(parameterService.find(Key.API_PRIMARY_OWNER_MODE, ParameterReferenceType.ENVIRONMENT)).thenReturn("HYBRID");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
-
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
-
-        verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember("user", null, MembershipMemberType.USER),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
     }
 
     @Test
-    public void shouldCreateImportApiWithOnlyDefinitionWithPrimaryOwnerGroup() throws IOException, TechnicalException {
-        URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition+primaryOwnerGroup.json");
-        String toBeImport = Resources.toString(url, Charsets.UTF_8);
-        ApiEntity apiEntity = new ApiEntity();
-        Api api = new Api();
-        api.setId(API_ID);
-        apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
-        UserEntity admin = new UserEntity();
-        admin.setId("admin");
-        admin.setSource(SOURCE);
-        admin.setSourceId(API_ID);
-        UserEntity user = new UserEntity();
-        user.setId("user");
-        user.setSource(SOURCE);
-        user.setSourceId(API_ID);
-
-        GroupEntity group = new GroupEntity();
-        group.setApiPrimaryOwner(user.getId());
-        group.setId("group");
-
-        when(groupService.findById(any())).thenReturn(group);
-        when(parameterService.find(Key.API_PRIMARY_OWNER_MODE, ParameterReferenceType.ENVIRONMENT)).thenReturn("GROUP");
-
-        apiService.createWithImportedDefinition(toBeImport, "admin");
-
-        verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
-
-        verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember("group", null, MembershipMemberType.GROUP),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
-    }
-
-    @Test
-    public void shouldCreateImportApiWithOnlyNewDefinition() throws IOException, TechnicalException {
+    public void shouldCreateImportApiWithOnlyNewDefinition() throws IOException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-new-api.definition.json");
         String toBeImport = Resources.toString(url, Charsets.UTF_8);
         ApiEntity apiEntity = new ApiEntity();
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
@@ -444,58 +300,46 @@ public class ApiService_CreateWithDefinitionTest {
         user.setId("user");
         user.setSource(SOURCE);
         user.setSourceId(API_ID);
-        when(userService.findById(admin.getId())).thenReturn(admin);
 
-        ApiEntity apiEntityCreated = apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
+
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
 
         verify(pageService, times(1)).createPage(any(), any(NewPageEntity.class));
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 
     @Test
-    public void shouldCreateImportApiWithOnlyDefinitionEnumLowerCase() throws IOException, TechnicalException {
+    public void shouldCreateImportApiWithOnlyDefinitionEnumLowerCase() throws IOException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition_enum_lowercase.json");
         String toBeImport = Resources.toString(url, Charsets.UTF_8);
         ApiEntity apiEntity = new ApiEntity();
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
         admin.setSourceId(API_ID);
-        when(userService.findById(admin.getId())).thenReturn(admin);
         when(planService.findById(anyString())).thenThrow(PlanNotFoundException.class);
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
         verify(pageService, times(1)).createPage(eq(API_ID), any(NewPageEntity.class));
-        verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember("admin", null, MembershipMemberType.USER),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 
     @Test
-    public void shouldCreateImportApiWithPlans() throws IOException, TechnicalException {
+    public void shouldCreateImportApiWithPlans() throws IOException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition+plans.json");
         String toBeImport = Resources.toString(url, Charsets.UTF_8);
         ApiEntity apiEntity = new ApiEntity();
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
@@ -504,47 +348,34 @@ public class ApiService_CreateWithDefinitionTest {
         user.setId("user");
         user.setSource(SOURCE);
         user.setSourceId(API_ID);
-        when(userService.findById(admin.getId())).thenReturn(admin);
         when(planService.findById(anyString())).thenThrow(PlanNotFoundException.class);
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
+
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
 
         verify(planService, times(2)).create(any(NewPlanEntity.class));
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 
     @Test
-    public void shouldCreateImportApiWithMetadata() throws IOException, TechnicalException {
+    public void shouldCreateImportApiWithMetadata() throws IOException {
         URL url = Resources.getResource("io/gravitee/rest/api/management/service/import-api.definition+metadata.json");
         String toBeImport = Resources.toString(url, Charsets.UTF_8);
         ApiEntity apiEntity = new ApiEntity();
         Api api = new Api();
         api.setId(API_ID);
         apiEntity.setId(API_ID);
-        when(apiRepository.findById(anyString())).thenReturn(Optional.empty());
-        when(apiRepository.create(any())).thenReturn(api);
+        when(apiService.createWithApiDefinition(any(), any(), any())).thenReturn(apiEntity);
+
         UserEntity admin = new UserEntity();
         admin.setId("admin");
         admin.setSource(SOURCE);
         admin.setSourceId(API_ID);
-        when(userService.findById(admin.getId())).thenReturn(admin);
 
-        apiService.createWithImportedDefinition(toBeImport, "admin");
+        apiDuplicatorService.createWithImportedDefinition(toBeImport, "admin");
 
-        verify(apiMetadataService, times(1)).create(any(NewApiMetadataEntity.class));
+        verify(apiService, times(1)).createWithApiDefinition(any(), eq("admin"), any());
+
         verify(apiMetadataService, times(2)).update(any(UpdateApiMetadataEntity.class));
-        verify(membershipService, times(1))
-            .addRoleToMemberOnReference(
-                new MembershipService.MembershipReference(MembershipReferenceType.API, API_ID),
-                new MembershipService.MembershipMember("admin", null, MembershipMemberType.USER),
-                new MembershipService.MembershipRole(RoleScope.API, SystemRole.PRIMARY_OWNER.name())
-            );
-        verify(apiRepository, never()).update(any());
-        verify(apiRepository, times(1)).create(any());
-        verify(genericNotificationConfigService, times(1)).create(any());
-        verify(searchEngineService, times(1)).index(any(), eq(false));
     }
 }
